ory integration:
1) Make AuthGuard session global and introduce IsPublic reflector
2) Pull type of authentication from config file
3) Merge Ory and Passport Login Forms into one page
