+++ server/auth/session.guard.ts
@@ -1,46 +1,57 @@
 import {
     CanActivate,
     ExecutionContext, Injectable,
-    UnauthorizedException,
 } from "@nestjs/common";
-import { Configuration, V0alpha2Api, V0alpha2ApiInterface } from '@ory/client';
-export class SessionGuard implements CanActivate {
-    constructor(private type: string = 'passport') {}
+import { Configuration, V0alpha2Api } from '@ory/client';
+import { Reflector } from "@nestjs/core";
+import {ConfigService} from "@nestjs/config";
 
+@Injectable()
+export class SessionGuard implements CanActivate {
+    constructor(
+        private configService: ConfigService,
+        private readonly reflector: Reflector
+    ) {}
+
+    // @ts-ignore
     async canActivate(context: ExecutionContext): Promise<boolean | Promise<boolean>> {
+        if (this.reflector.get<boolean>('public', context.getHandler())) {
+            return true
+        }
+
         const httpContext = context.switchToHttp();
         const request = httpContext.getRequest();
         const response = httpContext.getResponse();
-        if (this.type === 'ory') {
-            const ory = new V0alpha2Api(
-                new Configuration({
-                    basePath: process.env.ORY_SDK_URL,
-                    accessToken: process.env.ORY_ACCESS_TOKEN,
-                })
-            );
-            try {
+        const type = this.configService.get<string>('authentication_type')
+
+        try {
+            if (type === 'ory') {
+                const ory = new V0alpha2Api(
+                    new Configuration({
+                        basePath: this.configService.get<string>('ory.url'),
+                        accessToken: this.configService.get<string>('access_token'),
+                    })
+                );
                 const { data: session } = await ory.toSession(undefined, request.header('Cookie'))
-                request.user = {session};
+                request.user = { session };
                 return true
-            } catch (e) {
-                response.redirect("/ory-login");
             }
-        } else if (this.type === 'passport') {
-            try {
-                if (request.isAuthenticated && request.isAuthenticated()) {
-                    return true;
-                } else {
-                    if (request.url.startsWith("/api")) {
-                        return false;
-                    } else {
-                        response.redirect("/login");
-                    }
-                }
-            } catch (e) {
-                throw new UnauthorizedException();
+
+            if ((type === 'passport') && request.isAuthenticated && request.isAuthenticated()) {
+                return true;
             }
+            this.redirect(request, response);
+        } catch (e) {
+            // TODO: logging
+            this.redirect(request, response);
+        }
+    }
+
+    redirect(request, response) {
+        if (request.url.startsWith("/api")) {
+            return false;
         } else {
-            return false
+            response.redirect("/login");
         }
     }
 }
