+++ server/decorators/is-public.decorator.ts
@@ -1,0 +1,3 @@
+import { SetMetadata } from "@nestjs/common";
+
+export const IsPublic = () => SetMetadata('public', true)
