+++ server/ory/ory.controller.ts
@@ -10,6 +10,7 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import { parse } from "url";
 import { ViewService } from "../view/view.service";
 import OryService from "./ory.service";
+import {IsPublic} from "../decorators/is-public.decorator";
 
 @Controller()
 export default class OryController {
@@ -19,22 +20,7 @@ export default class OryController {
         private oryService: OryService
     ) {}
 
-    @Get("ory-login")
-    public async loginWithOry(@Req() req: Request, @Res() res: Response) {
-        const parsedUrl = parse(req.url, true);
-        // @ts-ignore
-        req.language = req.headers["accept-language"] || "en";
-
-        await this.viewService
-            .getNextServer()
-            .render(
-                req,
-                res,
-                "/ory-login-page",
-                Object.assign(parsedUrl.query)
-            );
-    }
-
+    @IsPublic()
     @Get("api/.ory/*")
     public async getOryPaths(@Req() req: NextApiRequest, @Res() res: NextApiResponse) {
         const parsedUrl = parse(req.url, true);
@@ -47,6 +33,7 @@ export default class OryController {
             )
     }
 
+    @IsPublic()
     @Post("api/.ory/*")
     public async oryPaths(@Req() req: NextApiRequest, @Res() res: NextApiResponse) {
         const parsedUrl = parse(req.url, true);
