+++ server/sitemap/sitemap.controller.ts
@@ -4,15 +4,15 @@ import {
     Header,
     Request,
     Res,
-    UseGuards,
 } from "@nestjs/common";
-import { SessionGuard } from "../auth/session.guard";
 import { SitemapService } from "./sitemap.service";
+import {IsPublic} from "../decorators/is-public.decorator";
 
 @Controller("/")
 export class SitemapController {
     constructor(private sitemapService: SitemapService) {}
 
+    @IsPublic()
     @Get("sitemap.xml")
     @Header("Content-Type", "text/xml")
     async getSitemap(@Res() res, @Request() req) {
@@ -22,7 +22,6 @@ export class SitemapController {
         res.send(sitemapXml);
     }
 
-    @UseGuards(SessionGuard)
     @Get("submit-sitemap")
     async submitSitemap(@Request() req) {
         return this.sitemapService.submitSitemap(
