+++ server/users/users.module.ts
@@ -5,6 +5,7 @@ import { UsersService } from "./users.service";
 import { UsersController } from "./users.controller";
 import { ViewModule } from "../view/view.module";
 import OryModule from "../ory/ory.module";
+import {ConfigModule} from "@nestjs/config";
 
 const UserModel = MongooseModule.forFeatureAsync([
     {
@@ -20,7 +21,7 @@ const UserModel = MongooseModule.forFeatureAsync([
 ]);
 
 @Module({
-    imports: [UserModel, ViewModule, OryModule],
+    imports: [UserModel, ViewModule, OryModule, ConfigModule],
     exports: [UsersService, UserModel],
     controllers: [UsersController],
     providers: [UsersService],
