+++ src/components/Login/LoginView.tsx
@@ -1,46 +1,11 @@
 import React, { useState } from "react";
-import { Form, Row, Col, Card, message, Typography } from "antd";
-import api from "../../api/user";
-import BackButton from "../BackButton";
-import CTARegistration from "../Home/CTARegistration";
-import { useTranslation } from "next-i18next";
-import { useRouter } from "next/router";
-import InputPassword from "../InputPassword";
-import Input from "../Input";
-import Button, { ButtonType } from "../Button";
+import { Row, Col, Card } from "antd";
 import colors from "../../styles/colors";
+import PassportLoginForm from "./PassportLoginForm";
+import OryLoginForm from "./OryLoginForm";
 
 const LoginView = (props) => {
-    const { t } = useTranslation();
-    const { previousUrl, host } = props
-    const router = useRouter();
     const [formType, setFormType] = useState("login");
-    const shouldGoBack = previousUrl.startsWith(host)
-
-    const onFinishFailed = errorInfo => {
-        if (typeof errorInfo === "string") {
-            message.error(errorInfo);
-        } else {
-            message.error(t("login:loginFailedMessage"));
-        }
-    };
-
-    const onFinish = async values => {
-        const result = await api.login(values, t);
-
-        if (!result.login) {
-            onFinishFailed(result.message);
-        } else {
-            message.success(t("login:loginSuccessfulMessage"));
-            if (shouldGoBack) {
-                router.back()
-            }
-            else {
-                router.push('/home')
-            }
-        }
-    };
-
     return (
         <>
             <Row justify="center">
@@ -53,89 +18,10 @@ const LoginView = (props) => {
                             })
                         }}
                     >
-                        {formType === "login" && (
-                            <>
-                                <Row className="typo-grey typo-center">
-                                    <Typography.Title style={{ fontSize: '1.5em', fontWeight: 500 }}>{t("login:formHeader")}</Typography.Title>
-                                </Row>
-                                <Form
-                                    name="basic"
-                                    initialValues={{ remember: true }}
-                                    onFinish={onFinish}
-                                    onFinishFailed={onFinishFailed}
-                                >
-                                    <Form.Item
-                                        label={t("login:emailLabel")}
-                                        name="email"
-                                        rules={[
-                                            {
-                                                required: true,
-                                                message: t(
-                                                    "login:emailErrorMessage"
-                                                )
-                                            }
-                                        ]}
-                                    >
-                                        <Input />
-                                    </Form.Item>
-
-                                    <Form.Item
-                                        label={t("login:passwordLabel")}
-                                        name="password"
-                                        rules={[
-                                            {
-                                                required: true,
-                                                message: t(
-                                                    "login:passwordErrorMessage"
-                                                )
-                                            }
-                                        ]}
-                                    >
-                                        <InputPassword />
-                                    </Form.Item>
-                                    <Form.Item>
-                                        <div
-                                            style={{
-                                                justifyContent:
-                                                    "space-between",
-                                                display: "flex"
-                                            }}
-                                        >
-                                            <a
-                                                onClick={() => {
-                                                    setFormType("signup");
-                                                }}
-                                                style={{
-                                                    color: colors.bluePrimary
-                                                }}
-                                            >
-                                                {t("login:signup")}
-                                            </a>
-                                            <Button
-                                                type={ButtonType.blue}
-                                                htmlType="submit"
-                                            >
-                                                {t("login:submitButton")}
-                                            </Button>
-                                        </div>
-                                    </Form.Item>
-                                </Form>
-                            </>
-                        )}
-                        {formType === "signup" && (
-                            <>
-                                <BackButton
-                                    callback={() =>
-                                        setFormType("login")
-                                    }
-                                    style={{
-                                        color: "#fff",
-                                        textDecoration: "underline"
-                                    }}
-                                />
-                                <CTARegistration />
-                            </>
-                        )}
+                        {props.type === 'ory'
+                            ? <OryLoginForm />
+                            : <PassportLoginForm {...props} formType={formType} setFormType={setFormType} />
+                        }
                     </Card>
                 </Col>
             </Row>
