+++ src/pages/login.tsx
@@ -15,13 +15,14 @@ const LoginPage: NextPage<{ data: any }> = (props) => {
     );
 };
 
-export async function getServerSideProps({ locale, locales, req }) {
+export async function getServerSideProps({ locale, locales, req, query }) {
     locale = parser.pick(locales, req.language) || locale || "en";
     return {
         props: {
             ...(await serverSideTranslations(locale)),
-            previousUrl: req.headers.referer || "none",
-            host: req.protocol + "://" + req.get("host"),
+            previousUrl: req.headers.referer || 'none',
+            host: req.protocol + '://' + req.get('host'),
+            type: query?.authType
         },
     };
 }
